Implementation of cRBM as real python project.
Now that the basic implementation of the convolutional RBM seems
to work, it has deserved its place as real program to not spam
the notebook with code.
Furthermore, the training procedure now takes into account
the double stranded structure of DNA (but this slows things down).
And the training works with the data as shared vars (givens).