New saving method, refactored parameter setting.
Saving now also saves the observers, making it possible to retrieve their scores
only afterwards.
Further, a new observer was added, the motifObserver. This class
protocols changes in the motifs over epochs.

The whole hyper parameter setting system was refactored
in a way such that it is possible to easily add new parameters.